+++ CHICCHIC/src/components/Navbar.tsx
@@ -1,48 +1,15 @@
-import { useState, useRef, useEffect } from "react";
+import { useState } from "react";
 import { useNavigate } from "react-router-dom";
 import { Menu, X } from "lucide-react";
-import { searchProductByName } from "../apis/productSearch";
-import type {ProductSearchResult } from "../types/searchtype";
 import mainlogo from "../assets/icons/main-logo.svg";
 import myhomelogo from "../assets/icons/myhome-logo.svg";
 import scraplogo from "../assets/icons/scrap-logo.svg";
 import { navbarMenu } from "../mocks/navbarMenu"
+import SearchNavbar from "./SearchNavbar"; // 검색창 + 드롭다운 컴포넌트
 
 const Navbar = () => {
-  const [searchQuery, setSearchQuery] = useState("");
-  const [searchResult, setSearchResult] = useState<ProductSearchResult | null>(null);
-  const [showDropdown, setShowDropdown] = useState(false);
   const [menuOpen, setMenuOpen] = useState(false);
   const navigate = useNavigate();
-  const dropdownRef = useRef<HTMLDivElement>(null);
-
-  // 검색어 변경 시 API 호출
-  useEffect(() => {
-    const fetch = async () => {
-      if (searchQuery.trim().length === 0) {
-        setSearchResult(null);
-        setShowDropdown(false);
-        return;
-      }
-      const result = await searchProductByName(searchQuery.trim());
-      setSearchResult(result);
-      setShowDropdown(!!result);
-    };
-    const timer = setTimeout(fetch, 300);
-    return () => clearTimeout(timer);
-  }, [searchQuery]);
-
-  useEffect(() => {
-    const handleClick = (e: MouseEvent) => {
-      if (dropdownRef.current && !dropdownRef.current.contains(e.target as Node)) {
-        setShowDropdown(false);
-      }
-    };
-    if (showDropdown) {
-      document.addEventListener("mousedown", handleClick);
-    }
-    return () => document.removeEventListener("mousedown", handleClick);
-  }, [showDropdown]);
 
   const handleNavigate = (path: string) => {
     navigate(path);
@@ -75,92 +42,7 @@ const Navbar = () => {
         </ul>
 
         <div className="flex-grow"></div>
-
-        <div className="relative hidden md:block mr-2 sm:mr-4" ref={dropdownRef}>
-          <input
-            type="text"
-            placeholder="Search"
-            value={searchQuery}
-            onChange={(e) => setSearchQuery(e.target.value)}
-            className="px-3 sm:px-4 py-1 text-xs sm:text-sm border rounded-3xl w-32 sm:w-48 lg:w-56 xl:w-72 focus:outline-none bg-transparent text-[#F7F4EF] border-[#F7F4EF]"
-            onFocus={() => { if (searchQuery.trim().length > 0) setShowDropdown(true); }}
-            autoComplete="off"
-            onKeyDown={async (e) => {
-              if (e.key === "Enter") {
-                const result = await searchProductByName(searchQuery.trim());
-                setSearchResult(result);
-                setShowDropdown(true);
-                if (result && result.name.trim().toLowerCase() === searchQuery.trim().toLowerCase()) {
-                  navigate(`/shopping/${result.id}`);
-                  setShowDropdown(false);
-                  setSearchQuery("");
-                }
-              }
-            }}
-          />
-          <button
-            className="absolute transform -translate-y-1/2 right-2 sm:right-3 top-1/2"
-            onClick={async () => {
-              const result = await searchProductByName(searchQuery.trim());
-              setSearchResult(result);
-              setShowDropdown(true);
-              if (result && result.name.trim().toLowerCase() === searchQuery.trim().toLowerCase()) {
-                navigate(`/shopping/${result.id}`);
-                setShowDropdown(false);
-                setSearchQuery("");
-              }
-            }}
-            tabIndex={-1}
-            type="button"
-          >
-            <svg
-              className="w-3 sm:w-4 h-3 cursor-pointer"
-              fill="none"
-              stroke="currentColor"
-              viewBox="0 0 24 24"
-            >
-              <path
-                strokeLinecap="round"
-                strokeLinejoin="round"
-                strokeWidth={2}
-                d="M21 21l-6-6m2-5a7 7 0 11-14 0 7 7 0 0114 0z"
-              />
-            </svg>
-          </button>
-
-          {/* 검색창 드롭다운 */}
-          {showDropdown && (
-            <div className="absolute left-0 mt-2 w-full bg-white text-[#AB3130] rounded-lg shadow-lg border border-[#AB3130]/20 z-50">
-              {searchResult ? (
-                <button
-                  className="w-full flex items-center gap-3 px-4 py-3 hover:bg-[#AB3130]/10 transition-colors text-left rounded-lg cursor-pointer"
-                  onClick={() => {
-                    navigate(`/shopping/${searchResult.id}`);
-                    setShowDropdown(false);
-                    setSearchQuery("");
-                  }}
-                >
-                  <img
-                    src={searchResult.imageUrl}
-                    alt={searchResult.name}
-                    className="w-10 h-10 object-cover rounded"
-                    onError={(e) => {
-                      e.currentTarget.src = "https://dummyimage.com/100x100/ccc/fff&text=No+Image";
-                    }}
-                  />
-                  <div>
-                    <div className="font-semibold">{searchResult.name}</div>
-                    <div className="text-xs text-[#AB3130]/70">{searchResult.brand} · {searchResult.ml}ml</div>
-                  </div>
-                </button>
-              ) : (
-                <div className="px-4 py-3 text-sm text-[#AB3130]/60 text-center select-none">
-                  검색 결과 없음
-                </div>
-              )}
-            </div>
-          )}
-        </div>
+        <SearchNavbar />
 
         <div className="flex items-center gap-2 sm:gap-4 flex-shrink-0">
           <button onClick={() => handleNavigate('/mypage/scraps')} className="hover:text-red-200 cursor-pointer">
